+++ sdk/src/androidTest/java/org/coralibre/android/sdk/internal/crypto/ENNumberTests.java
@@ -2,7 +2,7 @@ package org.coralibre.android.sdk.internal.crypto;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
+import org.coralibre.android.sdk.internal.crypto.ENNumber;
 import org.junit.Test;
 import org.junit.runner.RunWith;
