+++ sdk/src/androidTest/java/org/coralibre/android/sdk/internal/crypto/ExposeCheckerTests.java
@@ -4,16 +4,16 @@ import android.util.Pair;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
-import org.coralibre.android.sdk.internal.crypto.ppcp.ExposeChecker;
-import org.coralibre.android.sdk.internal.crypto.ppcp.RollingProximityIdentifier;
-import org.coralibre.android.sdk.internal.crypto.ppcp.TemporaryExposureKey;
+import org.coralibre.android.sdk.internal.crypto.ENNumber;
+import org.coralibre.android.sdk.internal.crypto.ExposeChecker;
+import org.coralibre.android.sdk.internal.crypto.RollingProximityIdentifier;
+import org.coralibre.android.sdk.internal.crypto.TemporaryExposureKey;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.coralibre.android.sdk.internal.crypto.ppcp.RollingProximityIdentifier.RPI_LENGTH;
-import static org.coralibre.android.sdk.internal.crypto.ppcp.TemporaryExposureKey.TEK_LENGTH;
-import static org.coralibre.android.sdk.internal.crypto.ppcp.TemporaryExposureKey.TEK_ROLLING_PERIOD;
+import static org.coralibre.android.sdk.internal.crypto.RollingProximityIdentifier.RPI_LENGTH;
+import static org.coralibre.android.sdk.internal.crypto.TemporaryExposureKey.TEK_LENGTH;
+import static org.coralibre.android.sdk.internal.crypto.TemporaryExposureKey.TEK_ROLLING_PERIOD;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
