+++ sdk/src/androidTest/java/org/coralibre/android/sdk/internal/crypto/PaddingDataTests.java
@@ -2,8 +2,8 @@ package org.coralibre.android.sdk.internal.crypto;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
-import org.coralibre.android.sdk.internal.crypto.ppcp.PaddedData;
+import org.coralibre.android.sdk.internal.crypto.ENNumber;
+import org.coralibre.android.sdk.internal.crypto.PaddedData;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import static org.junit.Assert.assertArrayEquals;
