+++ sdk/src/androidTest/java/org/coralibre/android/sdk/internal/crypto/TemporaryExposureKeyTests.java
@@ -2,8 +2,8 @@ package org.coralibre.android.sdk.internal.crypto;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
-import org.coralibre.android.sdk.internal.crypto.ppcp.TemporaryExposureKey;
+import org.coralibre.android.sdk.internal.crypto.ENNumber;
+import org.coralibre.android.sdk.internal.crypto.TemporaryExposureKey;
 import org.junit.Test;
 import org.junit.runner.RunWith;
