+++ sdk/src/main/java/org/coralibre/android/sdk/internal/crypto/AssociatedEncryptedMetadataKey.java
@@ -1,4 +1,4 @@
-package org.coralibre.android.sdk.internal.crypto.ppcp;
+package org.coralibre.android.sdk.internal.crypto;
 
 import java.security.InvalidParameterException;
