+++ sdk/src/main/java/org/coralibre/android/sdk/internal/crypto/CryptoException.java
@@ -1,4 +1,4 @@
-package org.coralibre.android.sdk.internal.crypto.ppcp;
+package org.coralibre.android.sdk.internal.crypto;
 
 public class CryptoException extends RuntimeException {
     CryptoException(Throwable t) {
