+++ sdk/src/main/java/org/coralibre/android/sdk/internal/crypto/ENNumber.java
@@ -1,4 +1,4 @@
-package org.coralibre.android.sdk.internal.crypto.ppcp;
+package org.coralibre.android.sdk.internal.crypto;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
