+++ sdk/src/main/java/org/coralibre/android/sdk/internal/crypto/RawTeKList.java
@@ -1,4 +1,4 @@
-package org.coralibre.android.sdk.internal.crypto.ppcp;
+package org.coralibre.android.sdk.internal.crypto;
 
 import android.util.Pair;
 import java.util.ArrayList;
