+++ sdk/src/main/java/org/coralibre/android/sdk/internal/database/Database.java
@@ -1,7 +1,7 @@
 package org.coralibre.android.sdk.internal.database;
 
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
+import org.coralibre.android.sdk.internal.crypto.ENNumber;
 import org.coralibre.android.sdk.internal.database.model.CapturedData;
 import org.coralibre.android.sdk.internal.database.model.GeneratedTEK;
 import org.coralibre.android.sdk.internal.database.model.IntervalOfCapturedData;
