+++ sdk/src/main/java/org/coralibre/android/sdk/internal/database/model/CapturedData.java
@@ -4,7 +4,7 @@ package org.coralibre.android.sdk.internal.database.model;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
+import org.coralibre.android.sdk.internal.crypto.ENNumber;
 
 public class CapturedData {
