+++ sdk/src/main/java/org/coralibre/android/sdk/internal/database/model/entity/EntityCapturedData.java
@@ -4,7 +4,7 @@ package org.coralibre.android.sdk.internal.database.model.entity;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
+import org.coralibre.android.sdk.internal.crypto.ENNumber;
 import org.coralibre.android.sdk.internal.database.model.CapturedData;
