+++ sdk/src/main/java/org/coralibre/android/sdk/internal/database/model/entity/EntityGeneratedTEK.java
@@ -3,7 +3,7 @@ package org.coralibre.android.sdk.internal.database.model.entity;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
+import org.coralibre.android.sdk.internal.crypto.ENNumber;
 import org.coralibre.android.sdk.internal.database.model.CapturedData;
 import org.coralibre.android.sdk.internal.database.model.GeneratedTEK;
