+++ sdk/src/main/java/org/coralibre/android/sdk/internal/database/persistent/ENNumberConverter.java
@@ -2,7 +2,7 @@ package org.coralibre.android.sdk.internal.database.persistent;
 
 import androidx.room.TypeConverter;
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
+import org.coralibre.android.sdk.internal.crypto.ENNumber;
 
 public class ENNumberConverter {
